Removed:
> The Metadata object MUST NOT be inserted in the Resv message by any
> implementation that conforms with this specification only.

Felt it was unnecessary and potentially confusing. Should be fine without.
(Especially given how we otherwise do not mandate how to fill in the
 Path message with metadata, but we equally do define how to process
 both Path and Resv - we explicitly say how a device compliant with this
spec just passes it on but takes noa ction against it.

modifications:

Moved the explanation of Flowspec=0 upfront into a section called
semantic of the metadata because i felt it provides a better reading.

Explained in more detail the difference between flowspec=0 and not
using the terminology of the framework (informative/advisory vs RSVP reservation).

Explained in that section then the additional options through
the new encoding and how the specific processing rules for those are not yet defined.

Gave the two encodings unique names and subsections (previously used the
same names).

I added explanation of the key differences between the encoding.

Minor word changes. Do a diff.



git-svn-id: http://10.51.54.244/svn/rfc@31 c7fd4c48-094c-430f-a73e-e8823c8a4d92